requirements.txt
RPi.GPIO
git+https://github.com/sn4k3/FakeRPi@master
